+++ src/main/java/ee/tuleva/onboarding/savings/fund/redemption/RedemptionVerificationService.java
@@ -6,8 +6,8 @@ import static ee.tuleva.onboarding.savings.SavingsFundOnboardingStatus.PENDING;
 import static ee.tuleva.onboarding.savings.fund.redemption.RedemptionRequest.Status.IN_REVIEW;
 import static ee.tuleva.onboarding.savings.fund.redemption.RedemptionRequest.Status.VERIFIED;
 
-import ee.tuleva.onboarding.aml.AmlService;
 import ee.tuleva.onboarding.aml.RiskLevels;
+import ee.tuleva.onboarding.aml.SanctionAndPepScreener;
 import ee.tuleva.onboarding.country.Country;
 import ee.tuleva.onboarding.kyb.LegalEntityScreener;
 import ee.tuleva.onboarding.kyc.KycCountryService;
@@ -30,7 +30,7 @@ public class RedemptionVerificationService {
   private final RedemptionStatusService redemptionStatusService;
   private final UserService userService;
   private final KycCountryService kycCountryService;
-  private final AmlService amlService;
+  private final SanctionAndPepScreener sanctionAndPepScreener;
   private final RiskLevels riskLevels;
   private final SavingsFundOnboardingRepository savingsFundOnboardingRepository;
   private final LegalEntityScreener legalEntityScreener;
@@ -89,9 +89,9 @@ public class RedemptionVerificationService {
                         "KYC survey with country not found: userId=" + user.getIdOrThrow()));
 
     Set<Country> allCountries = new HashSet<>(countries);
-    allCountries.addAll(amlService.recordedCitizenships(user));
+    allCountries.addAll(sanctionAndPepScreener.recordedCitizenships(user));
 
-    boolean screeningClear = amlService.isSanctionAndPepClear(user, allCountries);
+    boolean screeningClear = sanctionAndPepScreener.isSanctionAndPepClear(user, allCountries);
     boolean highRisk = riskLevels.isHighRisk(user.getPersonalCode());
     if (highRisk) {
       log.info(
