+++ src/test/groovy/ee/tuleva/onboarding/aml/AmlServiceTest.java
@@ -7,38 +7,24 @@ import static ee.tuleva.onboarding.kyc.KycCheck.RiskLevel.LOW;
 import static ee.tuleva.onboarding.mandate.MandateFixture.*;
 import static java.util.Arrays.stream;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.tuple;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anySet;
-import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.*;
 import static org.mockito.quality.Strictness.LENIENT;
 
 import ee.tuleva.onboarding.aml.notification.AmlCheckCreatedEvent;
-import ee.tuleva.onboarding.aml.notification.AmlChecksRunEvent;
-import ee.tuleva.onboarding.aml.sanctions.MatchResponse;
-import ee.tuleva.onboarding.aml.sanctions.PepAndSanctionCheckService;
-import ee.tuleva.onboarding.analytics.RecentThirdPillarCustomer;
-import ee.tuleva.onboarding.analytics.ThirdPillarAnalytics;
 import ee.tuleva.onboarding.auth.principal.Person;
 import ee.tuleva.onboarding.auth.principal.PersonImpl;
 import ee.tuleva.onboarding.conversion.ConversionResponse;
 import ee.tuleva.onboarding.conversion.UserConversionService;
-import ee.tuleva.onboarding.country.Countries;
-import ee.tuleva.onboarding.country.Country;
 import ee.tuleva.onboarding.epis.ContactDetails;
 import ee.tuleva.onboarding.event.TrackableEvent;
 import ee.tuleva.onboarding.kyc.KycCheck;
-import ee.tuleva.onboarding.kyc.KycCountryService;
 import ee.tuleva.onboarding.mandate.Mandate;
-import ee.tuleva.onboarding.notification.OperationsNotificationService;
 import ee.tuleva.onboarding.time.ClockHolder;
 import ee.tuleva.onboarding.user.User;
-import ee.tuleva.onboarding.user.UserRepository;
-import io.micrometer.core.instrument.MeterRegistry;
-import io.micrometer.core.instrument.simple.SimpleMeterRegistry;
 import java.time.Clock;
 import java.time.Instant;
 import java.time.ZoneOffset;
@@ -46,7 +32,6 @@ import java.time.temporal.ChronoUnit;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -59,15 +44,9 @@ import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.mockito.junit.jupiter.MockitoSettings;
 import org.springframework.context.ApplicationEventPublisher;
-import tools.jackson.databind.JsonNode;
-import tools.jackson.databind.json.JsonMapper;
-import tools.jackson.databind.node.ArrayNode;
-import tools.jackson.databind.node.JsonNodeFactory;
-import tools.jackson.databind.node.ObjectNode;
 
 @MockitoSettings(strictness = LENIENT)
 @ExtendWith(MockitoExtension.class)
@@ -75,22 +54,13 @@ class AmlServiceTest {
 
   @Mock private AmlCheckRepository amlCheckRepository;
   @Mock private ApplicationEventPublisher eventPublisher;
-  @Mock private PepAndSanctionCheckService pepAndSanctionCheckService;
-  @Mock private ThirdPillarAnalytics thirdPillarAnalytics;
-  @Mock private SavingsFundCustomers savingsFundCustomers;
-  @Mock private UserRepository userRepository;
   @Mock private UserConversionService userConversionService;
-  @Spy private JsonMapper jsonMapper = JsonMapper.builder().build();
-  @Spy private MeterRegistry meterRegistry = new SimpleMeterRegistry();
-  @Mock private OperationsNotificationService notificationService;
-  @Mock private KycCountryService kycCountryService;
 
   @InjectMocks private AmlService amlService;
 
   @Captor private ArgumentCaptor<AmlCheck> amlCheckCaptor;
   @Captor private ArgumentCaptor<TrackableEvent> trackableEventCaptor;
 
-  private static final JsonMapper objectMapper = JsonMapper.builder().build();
   private static final Instant FIXED_INSTANT = Instant.parse("2020-11-23T10:00:00Z");
   private static final Clock FIXED_CLOCK = Clock.fixed(FIXED_INSTANT, ZoneOffset.UTC);
   private Instant aYearAgoFromTestClock;
@@ -619,381 +589,6 @@ class AmlServiceTest {
     assertEquals(expectedResult.checksRequired, actualResult);
   }
 
-  @Test
-  void addSanctionAndPepCheckIfMissing_addsChecks() {
-    // given
-    User user = createUser("123", "First", "Last", 1L);
-    Set<Country> country = Countries.of("EE");
-    ArrayNode results = objectMapper.createArrayNode();
-    ObjectNode result1 = objectMapper.createObjectNode();
-    result1.put("id", "res123");
-    result1.put("match", true);
-    ArrayNode topics = objectMapper.createArrayNode();
-    topics.add("role.pep");
-    ObjectNode properties = objectMapper.createObjectNode();
-    properties.set("topics", topics);
-    result1.set("properties", properties);
-    results.add(result1);
-    JsonNode query = objectMapper.createObjectNode();
-    MatchResponse matchResponse = new MatchResponse(results, query);
-
-    when(pepAndSanctionCheckService.match(user, country)).thenReturn(matchResponse);
-    when(amlCheckRepository.existsByPersonalCodeAndTypeAndCreatedTimeAfter(
-            anyString(), any(AmlCheckType.class), any(Instant.class)))
-        .thenReturn(false);
-    when(amlCheckRepository.findAllByPersonalCodeAndTypeAndSuccess(
-            user.getPersonalCode(), POLITICALLY_EXPOSED_PERSON_OVERRIDE, true))
-        .thenReturn(List.of());
-    when(amlCheckRepository.findAllByPersonalCodeAndTypeAndSuccess(
-            user.getPersonalCode(), SANCTION_OVERRIDE, true))
-        .thenReturn(List.of());
-
-    // when
-    List<AmlCheck> addedChecks = amlService.addSanctionAndPepCheckIfMissing(user, country);
-
-    // then
-    verify(amlCheckRepository, times(2)).save(amlCheckCaptor.capture());
-    List<AmlCheck> savedChecks = amlCheckCaptor.getAllValues();
-
-    AmlCheck pepCheck =
-        savedChecks.stream()
-            .filter(c -> c.getType() == POLITICALLY_EXPOSED_PERSON_AUTO)
-            .findFirst()
-            .orElseThrow();
-    AmlCheck sanctionCheck =
-        savedChecks.stream().filter(c -> c.getType() == SANCTION).findFirst().orElseThrow();
-
-    assertFalse(pepCheck.isSuccess(), "PEP check should fail due to match");
-    assertTrue(sanctionCheck.isSuccess(), "Sanction check should pass as no sanction topic match");
-
-    assertEquals(2, addedChecks.size());
-    assertTrue(addedChecks.contains(pepCheck));
-    assertTrue(addedChecks.contains(sanctionCheck));
-
-    verify(eventPublisher, times(2)).publishEvent(any(AmlCheckCreatedEvent.class));
-  }
-
-  @Test
-  void addSanctionAndPepCheckIfMissing_considersOverrides() {
-    // given
-    User user = createUser("123", "First", "Last", 1L);
-    Set<Country> country = Countries.of("EE");
-
-    ArrayNode resultsArray = objectMapper.createArrayNode();
-    ObjectNode resultNode = JsonNodeFactory.instance.objectNode();
-    resultNode.put("id", "matchId123");
-    resultNode.put("match", true);
-    ObjectNode propertiesNode = JsonNodeFactory.instance.objectNode();
-    propertiesNode.set(
-        "topics", JsonNodeFactory.instance.arrayNode().add("role.pep").add("sanction"));
-    resultNode.set("properties", propertiesNode);
-    resultsArray.add(resultNode);
-    JsonNode queryNode = JsonNodeFactory.instance.objectNode();
-    MatchResponse matchResponse = new MatchResponse(resultsArray, queryNode);
-
-    when(pepAndSanctionCheckService.match(user, country)).thenReturn(matchResponse);
-    when(amlCheckRepository.existsByPersonalCodeAndTypeAndCreatedTimeAfter(
-            anyString(), any(AmlCheckType.class), any(Instant.class)))
-        .thenReturn(false);
-
-    AmlCheck pepOverride =
-        AmlCheck.builder()
-            .type(POLITICALLY_EXPOSED_PERSON_OVERRIDE)
-            .success(true)
-            .metadata(Map.of("results", List.of(Map.of("id", "matchId123"))))
-            .build();
-    when(amlCheckRepository.findAllByPersonalCodeAndTypeAndSuccess(
-            user.getPersonalCode(), POLITICALLY_EXPOSED_PERSON_OVERRIDE, true))
-        .thenReturn(List.of(pepOverride));
-
-    when(amlCheckRepository.findAllByPersonalCodeAndTypeAndSuccess(
-            user.getPersonalCode(), SANCTION_OVERRIDE, true))
-        .thenReturn(List.of());
-
-    // when
-    List<AmlCheck> addedChecks = amlService.addSanctionAndPepCheckIfMissing(user, country);
-
-    // then
-    verify(amlCheckRepository, times(2)).save(amlCheckCaptor.capture());
-    List<AmlCheck> savedChecks = amlCheckCaptor.getAllValues();
-    AmlCheck pepAutoCheck =
-        savedChecks.stream()
-            .filter(c -> c.getType() == POLITICALLY_EXPOSED_PERSON_AUTO)
-            .findFirst()
-            .orElseThrow();
-    AmlCheck sanctionCheck =
-        savedChecks.stream().filter(c -> c.getType() == SANCTION).findFirst().orElseThrow();
-
-    assertTrue(pepAutoCheck.isSuccess(), "PEP check should be successful due to override");
-    assertFalse(
-        sanctionCheck.isSuccess(), "Sanction check should fail as there's a match and no override");
-  }
-
-  @Test
-  void blanketOverrideWithoutResultsMetadata_appliesToAllMatches() {
-    // given
-    User user = createUser("123", "First", "Last", 1L);
-    Set<Country> country = Countries.of("EE");
-
-    ArrayNode resultsArray = objectMapper.createArrayNode();
-    ObjectNode resultNode = JsonNodeFactory.instance.objectNode();
-    resultNode.put("id", "matchId123");
-    resultNode.put("match", true);
-    ObjectNode propertiesNode = JsonNodeFactory.instance.objectNode();
-    propertiesNode.set("topics", JsonNodeFactory.instance.arrayNode().add("role.pep"));
-    resultNode.set("properties", propertiesNode);
-    resultsArray.add(resultNode);
-    JsonNode queryNode = JsonNodeFactory.instance.objectNode();
-    MatchResponse matchResponse = new MatchResponse(resultsArray, queryNode);
-
-    when(pepAndSanctionCheckService.match(user, country)).thenReturn(matchResponse);
-    when(amlCheckRepository.existsByPersonalCodeAndTypeAndCreatedTimeAfter(
-            anyString(), any(AmlCheckType.class), any(Instant.class)))
-        .thenReturn(false);
-
-    AmlCheck blanketOverrideWithoutResults =
-        AmlCheck.builder()
-            .type(POLITICALLY_EXPOSED_PERSON_OVERRIDE)
-            .success(true)
-            .metadata(Map.of("comment", "Kinnitan ärisuhte", "createdBy", "admin"))
-            .build();
-    when(amlCheckRepository.findAllByPersonalCodeAndTypeAndSuccess(
-            user.getPersonalCode(), POLITICALLY_EXPOSED_PERSON_OVERRIDE, true))
-        .thenReturn(List.of(blanketOverrideWithoutResults));
-
-    when(amlCheckRepository.findAllByPersonalCodeAndTypeAndSuccess(
-            user.getPersonalCode(), SANCTION_OVERRIDE, true))
-        .thenReturn(List.of());
-
-    // when
-    amlService.addSanctionAndPepCheckIfMissing(user, country);
-
-    // then
-    verify(amlCheckRepository, times(2)).save(amlCheckCaptor.capture());
-    List<AmlCheck> savedChecks = amlCheckCaptor.getAllValues();
-    AmlCheck pepAutoCheck =
-        savedChecks.stream()
-            .filter(c -> c.getType() == POLITICALLY_EXPOSED_PERSON_AUTO)
-            .findFirst()
-            .orElseThrow();
-
-    assertThat(pepAutoCheck.isSuccess()).isTrue();
-  }
-
-  @Test
-  void overrideWithObjectShapedResultsMetadata_doesNotCrashAndAppliesOverride() {
-    User user = createUser("123", "First", "Last", 1L);
-    Set<Country> country = Countries.of("EE");
-
-    ArrayNode resultsArray = objectMapper.createArrayNode();
-    ObjectNode resultNode = JsonNodeFactory.instance.objectNode();
-    resultNode.put("id", "matchId123");
-    resultNode.put("match", true);
-    ObjectNode propertiesNode = JsonNodeFactory.instance.objectNode();
-    propertiesNode.set("topics", JsonNodeFactory.instance.arrayNode().add("role.pep"));
-    resultNode.set("properties", propertiesNode);
-    resultsArray.add(resultNode);
-    MatchResponse matchResponse =
-        new MatchResponse(resultsArray, JsonNodeFactory.instance.objectNode());
-
-    when(pepAndSanctionCheckService.match(user, country)).thenReturn(matchResponse);
-    when(amlCheckRepository.existsByPersonalCodeAndTypeAndCreatedTimeAfter(
-            anyString(), any(AmlCheckType.class), any(Instant.class)))
-        .thenReturn(false);
-
-    AmlCheck corruptOverride =
-        AmlCheck.builder()
-            .type(POLITICALLY_EXPOSED_PERSON_OVERRIDE)
-            .success(true)
-            .metadata(Map.of("results", Map.of("nodeType", "ARRAY", "array", true)))
-            .build();
-    when(amlCheckRepository.findAllByPersonalCodeAndTypeAndSuccess(
-            user.getPersonalCode(), POLITICALLY_EXPOSED_PERSON_OVERRIDE, true))
-        .thenReturn(List.of(corruptOverride));
-    when(amlCheckRepository.findAllByPersonalCodeAndTypeAndSuccess(
-            user.getPersonalCode(), SANCTION_OVERRIDE, true))
-        .thenReturn(List.of());
-
-    List<AmlCheck> addedChecks = amlService.addSanctionAndPepCheckIfMissing(user, country);
-
-    AmlCheck pepAutoCheck =
-        addedChecks.stream()
-            .filter(check -> check.getType() == POLITICALLY_EXPOSED_PERSON_AUTO)
-            .findFirst()
-            .orElseThrow();
-    assertThat(pepAutoCheck.isSuccess()).isTrue();
-  }
-
-  @Test
-  void overrideWithNonMatchingResults_doesNotApplyAndCheckFails() {
-    User user = createUser("123", "First", "Last", 1L);
-    Set<Country> country = Countries.of("EE");
-
-    ArrayNode resultsArray = objectMapper.createArrayNode();
-    ObjectNode resultNode = JsonNodeFactory.instance.objectNode();
-    resultNode.put("id", "currentMatch");
-    resultNode.put("match", true);
-    ObjectNode propertiesNode = JsonNodeFactory.instance.objectNode();
-    propertiesNode.set("topics", JsonNodeFactory.instance.arrayNode().add("role.pep"));
-    resultNode.set("properties", propertiesNode);
-    resultsArray.add(resultNode);
-    MatchResponse matchResponse =
-        new MatchResponse(resultsArray, JsonNodeFactory.instance.objectNode());
-
-    when(pepAndSanctionCheckService.match(user, country)).thenReturn(matchResponse);
-    when(amlCheckRepository.existsByPersonalCodeAndTypeAndCreatedTimeAfter(
-            anyString(), any(AmlCheckType.class), any(Instant.class)))
-        .thenReturn(false);
-
-    AmlCheck nonMatchingOverride =
-        AmlCheck.builder()
-            .type(POLITICALLY_EXPOSED_PERSON_OVERRIDE)
-            .success(true)
-            .metadata(Map.of("results", List.of("not-a-map", Map.of("id", "otherMatch"))))
-            .build();
-    when(amlCheckRepository.findAllByPersonalCodeAndTypeAndSuccess(
-            user.getPersonalCode(), POLITICALLY_EXPOSED_PERSON_OVERRIDE, true))
-        .thenReturn(List.of(nonMatchingOverride));
-    when(amlCheckRepository.findAllByPersonalCodeAndTypeAndSuccess(
-            user.getPersonalCode(), SANCTION_OVERRIDE, true))
-        .thenReturn(List.of());
-
-    List<AmlCheck> addedChecks = amlService.addSanctionAndPepCheckIfMissing(user, country);
-
-    AmlCheck pepAutoCheck =
-        addedChecks.stream()
-            .filter(check -> check.getType() == POLITICALLY_EXPOSED_PERSON_AUTO)
-            .findFirst()
-            .orElseThrow();
-    assertThat(pepAutoCheck.isSuccess()).isFalse();
-  }
-
-  @Test
-  void addSanctionAndPepCheckIfMissing_handlesMatchServiceException() {
-    // given
-    User user = createUser("123", "First", "Last", 1L);
-    Set<Country> country = Countries.of("EE");
-    when(pepAndSanctionCheckService.match(user, country))
-        .thenThrow(new RuntimeException("Match service error"));
-
-    // when
-    List<AmlCheck> result = amlService.addSanctionAndPepCheckIfMissing(user, country);
-
-    // then
-    assertTrue(result.isEmpty(), "Should return an empty list on exception");
-    verify(amlCheckRepository, never()).save(any());
-  }
-
-  @Test
-  void addSanctionAndPepCheckIfMissing_screeningFailureIsObservable() {
-    // given
-    User user = createUser("38001010000", "First", "Last", 1L);
-    Set<Country> country = Countries.of("EE");
-    when(pepAndSanctionCheckService.match(user, country))
-        .thenThrow(new RuntimeException("Match service error"));
-
-    // when
-    List<AmlCheck> result = amlService.addSanctionAndPepCheckIfMissing(user, country);
-
-    // then
-    assertTrue(result.isEmpty(), "Should return an empty list on exception");
-    assertEquals(
-        1.0,
-        meterRegistry.counter("aml.screening.failure", "phase", "match").count(),
-        "Screening failure should increment the aml.screening.failure metric");
-    verify(notificationService, never()).sendMessage(anyString(), any());
-  }
-
-  @Test
-  void runAmlChecksOnThirdPillarCustomers_sendsSingleAggregatedAlertOnScreeningFailures() {
-    // given
-    RecentThirdPillarCustomer ok = thirdPillarRecord("ok", "EE");
-    RecentThirdPillarCustomer fail1 = thirdPillarRecord("fail1", "EE");
-    RecentThirdPillarCustomer fail2 = thirdPillarRecord("fail2", "EE");
-    when(thirdPillarAnalytics.recentCustomers()).thenReturn(List.of(fail1, ok, fail2));
-
-    MatchResponse emptyMatchResponse =
-        new MatchResponse(objectMapper.createArrayNode(), objectMapper.createObjectNode());
-    when(pepAndSanctionCheckService.match(eq(ok), anySet())).thenReturn(emptyMatchResponse);
-    when(pepAndSanctionCheckService.match(eq(fail1), anySet()))
-        .thenThrow(new RuntimeException("Match service error"));
-    when(pepAndSanctionCheckService.match(eq(fail2), anySet()))
-        .thenThrow(new RuntimeException("Match service error"));
-    when(amlCheckRepository.existsByPersonalCodeAndTypeAndCreatedTimeAfter(
-            anyString(), any(AmlCheckType.class), any(Instant.class)))
-        .thenReturn(false);
-
-    // when
-    amlService.runAmlChecksOnThirdPillarCustomers();
-
-    // then
-    verify(notificationService, times(1))
-        .sendMessage(
-            "AML batch: sanction/PEP screening failed for 2 of 3 third-pillar customers this run",
-            OperationsNotificationService.Channel.AML);
-    verify(amlCheckRepository, times(2)).save(any(AmlCheck.class));
-    assertEquals(
-        2.0,
-        meterRegistry.counter("aml.screening.failure", "phase", "match").count(),
-        "Both screening failures should increment the metric");
-  }
-
-  @Test
-  void runAmlChecksOnThirdPillarCustomers_slackFailureDoesNotAbortBatch() {
-    // given
-    RecentThirdPillarCustomer c1 = thirdPillarRecord("c1", "EE");
-    RecentThirdPillarCustomer c2 = thirdPillarRecord("c2", "EE");
-    when(thirdPillarAnalytics.recentCustomers()).thenReturn(List.of(c1, c2));
-    when(pepAndSanctionCheckService.match(any(Person.class), anySet()))
-        .thenThrow(new RuntimeException("Match service error"));
-    doThrow(new RuntimeException("Slack is down"))
-        .when(notificationService)
-        .sendMessage(anyString(), any());
-
-    // when / then
-    assertDoesNotThrow(() -> amlService.runAmlChecksOnThirdPillarCustomers());
-
-    assertEquals(
-        2.0,
-        meterRegistry.counter("aml.screening.failure", "phase", "match").count(),
-        "Both customers should have been screened");
-  }
-
-  private RecentThirdPillarCustomer thirdPillarRecord(String personalCode, String country) {
-    return new RecentThirdPillarCustomer(personalCode, "First", "Last", country);
-  }
-
-  @Test
-  void runAmlChecksOnThirdPillarCustomers_processesRecords() {
-    // given
-    RecentThirdPillarCustomer record1 = new RecentThirdPillarCustomer("p1", "F1", "L1", "EE");
-    when(thirdPillarAnalytics.recentCustomers()).thenReturn(List.of(record1));
-
-    MatchResponse emptyMatchResponse =
-        new MatchResponse(objectMapper.createArrayNode(), objectMapper.createObjectNode());
-    when(pepAndSanctionCheckService.match(eq(record1), anySet())).thenReturn(emptyMatchResponse);
-    when(amlCheckRepository.existsByPersonalCodeAndTypeAndCreatedTimeAfter(
-            anyString(), any(AmlCheckType.class), any(Instant.class)))
-        .thenReturn(false);
-
-    when(amlCheckRepository.findAllByPersonalCodeAndTypeAndSuccess(
-            anyString(), eq(POLITICALLY_EXPOSED_PERSON_OVERRIDE), eq(true)))
-        .thenReturn(List.of());
-    when(amlCheckRepository.findAllByPersonalCodeAndTypeAndSuccess(
-            anyString(), eq(SANCTION_OVERRIDE), eq(true)))
-        .thenReturn(List.of());
-
-    // when
-    amlService.runAmlChecksOnThirdPillarCustomers();
-
-    // then
-    verify(eventPublisher).publishEvent(any(AmlChecksRunEvent.class));
-
-    verify(pepAndSanctionCheckService).match(record1, Countries.of("EE"));
-    verify(amlCheckRepository, times(2)).save(any(AmlCheck.class));
-    verify(eventPublisher, times(2)).publishEvent(any(AmlCheckCreatedEvent.class));
-  }
-
   private static List<AmlCheck> successfulChecks(AmlCheckType... checkTypes) {
     return stream(checkTypes)
         .map(type -> AmlCheck.builder().type(type).success(true).build())
@@ -1004,413 +599,5 @@ class AmlServiceTest {
     return stream(checkTypes)
         .map(type -> AmlCheck.builder().type(type).success(false).build())
         .toList();
-  }
-
-  private MatchResponse sanctionAndPepMatch() {
-    ArrayNode results = objectMapper.createArrayNode();
-    ObjectNode result = JsonNodeFactory.instance.objectNode();
-    result.put("id", "matchId123");
-    result.put("match", true);
-    ObjectNode properties = JsonNodeFactory.instance.objectNode();
-    properties.set("topics", JsonNodeFactory.instance.arrayNode().add("role.pep").add("sanction"));
-    result.set("properties", properties);
-    results.add(result);
-    return new MatchResponse(results, JsonNodeFactory.instance.objectNode());
-  }
-
-  private void latestCheckIs(AmlCheckType type, boolean success) {
-    when(amlCheckRepository.findFirstByPersonalCodeAndTypeOrderByCreatedTimeDescIdDesc(
-            anyString(), eq(type)))
-        .thenReturn(
-            Optional.of(
-                AmlCheck.builder()
-                    .personalCode("38812121215")
-                    .type(type)
-                    .success(success)
-                    .build()));
-  }
-
-  @Test
-  void screening_recordsNewlyFailingResultDespiteARecentPassingCheck() {
-    User user = createUser("38812121215", "Test", "User", 1L);
-    when(pepAndSanctionCheckService.match(eq(user), anySet())).thenReturn(sanctionAndPepMatch());
-    when(amlCheckRepository.existsByPersonalCodeAndTypeAndCreatedTimeAfter(
-            anyString(), any(AmlCheckType.class), any(Instant.class)))
-        .thenReturn(true);
-    latestCheckIs(SANCTION, true);
-    latestCheckIs(POLITICALLY_EXPOSED_PERSON_AUTO, true);
-
-    amlService.addSanctionAndPepCheckIfMissing(user, Countries.of("EE"));
-
-    verify(amlCheckRepository, times(2)).save(amlCheckCaptor.capture());
-    assertThat(amlCheckCaptor.getAllValues())
-        .extracting(AmlCheck::getType, AmlCheck::isSuccess)
-        .containsExactlyInAnyOrder(
-            tuple(POLITICALLY_EXPOSED_PERSON_AUTO, false), tuple(SANCTION, false));
-  }
-
-  @Test
-  void screening_recordsRecoveryWhenAPreviouslyFailingPersonNoLongerMatches() {
-    User user = createUser("38812121215", "Test", "User", 1L);
-    when(pepAndSanctionCheckService.match(eq(user), anySet()))
-        .thenReturn(
-            new MatchResponse(objectMapper.createArrayNode(), objectMapper.createObjectNode()));
-    when(amlCheckRepository.existsByPersonalCodeAndTypeAndCreatedTimeAfter(
-            anyString(), any(AmlCheckType.class), any(Instant.class)))
-        .thenReturn(true);
-    latestCheckIs(SANCTION, false);
-    latestCheckIs(POLITICALLY_EXPOSED_PERSON_AUTO, false);
-
-    amlService.addSanctionAndPepCheckIfMissing(user, Countries.of("EE"));
-
-    verify(amlCheckRepository, times(2)).save(amlCheckCaptor.capture());
-    assertThat(amlCheckCaptor.getAllValues())
-        .extracting(AmlCheck::getType, AmlCheck::isSuccess)
-        .containsExactlyInAnyOrder(
-            tuple(POLITICALLY_EXPOSED_PERSON_AUTO, true), tuple(SANCTION, true));
-  }
-
-  @Test
-  void screening_doesNotRecordAnUnchangedOutcomeWithinTheWindow() {
-    User user = createUser("38812121215", "Test", "User", 1L);
-    when(pepAndSanctionCheckService.match(eq(user), anySet()))
-        .thenReturn(
-            new MatchResponse(objectMapper.createArrayNode(), objectMapper.createObjectNode()));
-    when(amlCheckRepository.existsByPersonalCodeAndTypeAndCreatedTimeAfter(
-            anyString(), any(AmlCheckType.class), any(Instant.class)))
-        .thenReturn(true);
-    latestCheckIs(SANCTION, true);
-    latestCheckIs(POLITICALLY_EXPOSED_PERSON_AUTO, true);
-
-    amlService.addSanctionAndPepCheckIfMissing(user, Countries.of("EE"));
-
-    verify(amlCheckRepository, never()).save(any(AmlCheck.class));
-  }
-
-  @Test
-  void screening_recordsAnUnchangedOutcomeOnceTheWindowHasExpired() {
-    User user = createUser("38812121215", "Test", "User", 1L);
-    when(pepAndSanctionCheckService.match(eq(user), anySet()))
-        .thenReturn(
-            new MatchResponse(objectMapper.createArrayNode(), objectMapper.createObjectNode()));
-    when(amlCheckRepository.existsByPersonalCodeAndTypeAndCreatedTimeAfter(
-            anyString(), any(AmlCheckType.class), any(Instant.class)))
-        .thenReturn(false);
-    latestCheckIs(SANCTION, true);
-    latestCheckIs(POLITICALLY_EXPOSED_PERSON_AUTO, true);
-
-    amlService.addSanctionAndPepCheckIfMissing(user, Countries.of("EE"));
-
-    verify(amlCheckRepository, times(2)).save(any(AmlCheck.class));
-  }
-
-  private User savingsFundCustomer(String personalCode) {
-    return savingsFundCustomer(personalCode, null);
-  }
-
-  private User savingsFundCustomer(String personalCode, Long id) {
-    return User.builder()
-        .id(id)
-        .personalCode(personalCode)
-        .firstName("First" + personalCode)
-        .lastName("Last" + personalCode)
-        .build();
-  }
-
-  private void givenRecordedCitizenships(String personalCode, List<String> citizenships) {
-    given(
-            amlCheckRepository.findFirstByPersonalCodeAndTypeOrderByCreatedTimeDescIdDesc(
-                personalCode, CUSTODY_RIGHT))
-        .willReturn(
-            Optional.of(
-                AmlCheck.builder()
-                    .personalCode(personalCode)
-                    .type(CUSTODY_RIGHT)
-                    .success(true)
-                    .metadata(Map.of("citizenships", citizenships))
-                    .build()));
-  }
-
-  @Test
-  void runAmlChecksOnSavingsFundCustomers_screensEachCustomerOnEveryCountryTheyAreTiedTo() {
-    User adult = savingsFundCustomer("38888888881", 1L);
-    User child = savingsFundCustomer("61506150006", 2L);
-    given(savingsFundCustomers.personalCodes())
-        .willReturn(List.of(adult.getPersonalCode(), child.getPersonalCode()));
-    given(
-            userRepository.findAllByPersonalCodeIn(
-                List.of(adult.getPersonalCode(), child.getPersonalCode())))
-        .willReturn(List.of(adult, child));
-    given(kycCountryService.getCountries(1L)).willReturn(Optional.of(Countries.of("EE", "RU")));
-    given(kycCountryService.getCountries(2L)).willReturn(Optional.empty());
-    givenRecordedCitizenships(child.getPersonalCode(), List.of("UA"));
-    given(pepAndSanctionCheckService.match(any(Person.class), anySet()))
-        .willReturn(
-            new MatchResponse(objectMapper.createArrayNode(), objectMapper.createObjectNode()));
-
-    amlService.runAmlChecksOnSavingsFundCustomers();
-
-    verify(pepAndSanctionCheckService).match(adult, Countries.of("EE", "RU"));
-    verify(pepAndSanctionCheckService).match(child, Countries.of("UA"));
-  }
-
-  @Test
-  void addSanctionAndPepCheckIfMissing_screensOnKycSurveyCountriesAndRecordedCitizenships() {
-    User user = savingsFundCustomer("38812121215", 3L);
-    given(kycCountryService.getCountries(3L)).willReturn(Optional.of(Countries.of("EE", "RU")));
-    givenRecordedCitizenships(user.getPersonalCode(), List.of("UA"));
-    given(pepAndSanctionCheckService.match(any(Person.class), anySet()))
-        .willReturn(
-            new MatchResponse(objectMapper.createArrayNode(), objectMapper.createObjectNode()));
-
-    amlService.addSanctionAndPepCheckIfMissing(user);
-
-    verify(pepAndSanctionCheckService).match(user, Countries.of("EE", "RU", "UA"));
-  }
-
-  @Test
-  void addSanctionAndPepCheckIfMissing_screensOnEstoniaOnlyWhenNothingIsKnownAboutTheUser() {
-    User user = savingsFundCustomer("38812121215", 4L);
-    given(kycCountryService.getCountries(4L)).willReturn(Optional.empty());
-    given(pepAndSanctionCheckService.match(any(Person.class), anySet()))
-        .willReturn(
-            new MatchResponse(objectMapper.createArrayNode(), objectMapper.createObjectNode()));
-
-    amlService.addSanctionAndPepCheckIfMissing(user);
-
-    verify(pepAndSanctionCheckService).match(user, Countries.<String>of());
-  }
-
-  @Test
-  void runAmlChecksOnSavingsFundCustomers_screensEveryOnboardedPerson() {
-    User first = savingsFundCustomer("38888888881", 11L);
-    User second = savingsFundCustomer("38888888882", 12L);
-    when(savingsFundCustomers.personalCodes())
-        .thenReturn(List.of(first.getPersonalCode(), second.getPersonalCode()));
-    when(userRepository.findAllByPersonalCodeIn(
-            List.of(first.getPersonalCode(), second.getPersonalCode())))
-        .thenReturn(List.of(first, second));
-    when(pepAndSanctionCheckService.match(any(Person.class), anySet()))
-        .thenReturn(
-            new MatchResponse(objectMapper.createArrayNode(), objectMapper.createObjectNode()));
-    when(amlCheckRepository.existsByPersonalCodeAndTypeAndCreatedTimeAfter(
-            anyString(), any(AmlCheckType.class), any(Instant.class)))
-        .thenReturn(false);
-
-    amlService.runAmlChecksOnSavingsFundCustomers();
-
-    verify(pepAndSanctionCheckService).match(first, Countries.<String>of());
-    verify(pepAndSanctionCheckService).match(second, Countries.<String>of());
-    verify(notificationService, never()).sendMessage(anyString(), any());
-  }
-
-  @Test
-  void runAmlChecksOnSavingsFundCustomers_sendsSingleAggregatedAlertOnScreeningFailures() {
-    User ok = savingsFundCustomer("38888888881", 11L);
-    User failing = savingsFundCustomer("38888888882", 12L);
-    when(savingsFundCustomers.personalCodes())
-        .thenReturn(List.of(ok.getPersonalCode(), failing.getPersonalCode()));
-    when(userRepository.findAllByPersonalCodeIn(any())).thenReturn(List.of(ok, failing));
-    when(pepAndSanctionCheckService.match(eq(ok), anySet()))
-        .thenReturn(
-            new MatchResponse(objectMapper.createArrayNode(), objectMapper.createObjectNode()));
-    when(pepAndSanctionCheckService.match(eq(failing), anySet()))
-        .thenThrow(new RuntimeException("Match service error"));
-    when(amlCheckRepository.existsByPersonalCodeAndTypeAndCreatedTimeAfter(
-            anyString(), any(AmlCheckType.class), any(Instant.class)))
-        .thenReturn(false);
-
-    amlService.runAmlChecksOnSavingsFundCustomers();
-
-    verify(notificationService, times(1))
-        .sendMessage(
-            "AML batch: sanction/PEP screening failed for 1 of 2 savings fund customers this run",
-            OperationsNotificationService.Channel.AML);
-  }
-
-  @Test
-  void runAmlChecksOnSavingsFundCustomers_countsCheckPersistenceFailuresAndContinues() {
-    User failing = savingsFundCustomer("38888888881", 11L);
-    when(savingsFundCustomers.personalCodes()).thenReturn(List.of(failing.getPersonalCode()));
-    when(userRepository.findAllByPersonalCodeIn(any())).thenReturn(List.of(failing));
-    when(pepAndSanctionCheckService.match(any(Person.class), anySet()))
-        .thenReturn(
-            new MatchResponse(objectMapper.createArrayNode(), objectMapper.createObjectNode()));
-    when(amlCheckRepository.existsByPersonalCodeAndTypeAndCreatedTimeAfter(
-            anyString(), any(AmlCheckType.class), any(Instant.class)))
-        .thenThrow(new RuntimeException("Database unavailable"));
-
-    assertDoesNotThrow(() -> amlService.runAmlChecksOnSavingsFundCustomers());
-
-    verify(notificationService)
-        .sendMessage(
-            "AML batch: sanction/PEP screening failed for 1 of 1 savings fund customers this run",
-            OperationsNotificationService.Channel.AML);
-    assertEquals(
-        1.0, meterRegistry.counter("aml.screening.failure", "phase", "savings-fund-batch").count());
-  }
-
-  @Test
-  void runAmlChecksOnSavingsFundCustomers_slackFailureDoesNotAbortBatch() {
-    User failing = savingsFundCustomer("38888888881", 11L);
-    when(savingsFundCustomers.personalCodes()).thenReturn(List.of(failing.getPersonalCode()));
-    when(userRepository.findAllByPersonalCodeIn(any())).thenReturn(List.of(failing));
-    when(pepAndSanctionCheckService.match(any(Person.class), anySet()))
-        .thenThrow(new RuntimeException("Match service error"));
-    doThrow(new RuntimeException("Slack down"))
-        .when(notificationService)
-        .sendMessage(anyString(), any());
-
-    assertDoesNotThrow(() -> amlService.runAmlChecksOnSavingsFundCustomers());
-  }
-
-  @Test
-  void runAmlChecksOnSavingsFundCustomers_doesNothingWhenNobodyIsOnboarded() {
-    when(savingsFundCustomers.personalCodes()).thenReturn(List.of());
-    when(userRepository.findAllByPersonalCodeIn(List.of())).thenReturn(List.of());
-
-    amlService.runAmlChecksOnSavingsFundCustomers();
-
-    verify(pepAndSanctionCheckService, never()).match(any(Person.class), anySet());
-    verify(notificationService, never()).sendMessage(anyString(), any());
-  }
-
-  @Test
-  void isSanctionAndPepClear_failsClosedWhenScreeningThrows() {
-    User user = createUser("123", "First", "Last", 1L);
-    Set<Country> country = Countries.of("EE");
-    when(pepAndSanctionCheckService.match(user, country))
-        .thenThrow(new RuntimeException("screening service down"));
-
-    assertFalse(amlService.isSanctionAndPepClear(user, country));
-  }
-
-  @Test
-  void isSanctionAndPepClear_trueWhenLatestScreeningChecksPass() {
-    User user = createUser("123", "First", "Last", 1L);
-    Set<Country> country = Countries.of("EE");
-    MatchResponse emptyResponse =
-        new MatchResponse(objectMapper.createArrayNode(), objectMapper.createObjectNode());
-    when(pepAndSanctionCheckService.match(user, country)).thenReturn(emptyResponse);
-    when(amlCheckRepository.findFirstByPersonalCodeAndTypeOrderByCreatedTimeDescIdDesc(
-            "123", SANCTION))
-        .thenReturn(
-            Optional.of(
-                AmlCheck.builder().personalCode("123").type(SANCTION).success(true).build()));
-    when(amlCheckRepository.findFirstByPersonalCodeAndTypeOrderByCreatedTimeDescIdDesc(
-            "123", POLITICALLY_EXPOSED_PERSON_AUTO))
-        .thenReturn(
-            Optional.of(
-                AmlCheck.builder()
-                    .personalCode("123")
-                    .type(POLITICALLY_EXPOSED_PERSON_AUTO)
-                    .success(true)
-                    .build()));
-
-    assertTrue(amlService.isSanctionAndPepClear(user, country));
-  }
-
-  @Test
-  void isSanctionAndPepClear_falseWhenLatestSanctionCheckHasFailed() {
-    User user = createUser("123", "First", "Last", 1L);
-    Set<Country> country = Countries.of("EE");
-    ArrayNode results = objectMapper.createArrayNode();
-    ObjectNode result = objectMapper.createObjectNode();
-    result.put("id", "sanction123");
-    result.put("match", true);
-    ArrayNode topics = objectMapper.createArrayNode();
-    topics.add("sanction");
-    ObjectNode properties = objectMapper.createObjectNode();
-    properties.set("topics", topics);
-    result.set("properties", properties);
-    results.add(result);
-    MatchResponse matchResponse = new MatchResponse(results, objectMapper.createObjectNode());
-    when(pepAndSanctionCheckService.match(user, country)).thenReturn(matchResponse);
-    when(amlCheckRepository.findFirstByPersonalCodeAndTypeOrderByCreatedTimeDescIdDesc(
-            "123", SANCTION))
-        .thenReturn(
-            Optional.of(
-                AmlCheck.builder().personalCode("123").type(SANCTION).success(false).build()));
-
-    assertFalse(amlService.isSanctionAndPepClear(user, country));
-  }
-
-  @Test
-  void isSanctionAndPepClear_falseWhenNoScreeningRecordExists() {
-    User user = createUser("123", "First", "Last", 1L);
-    Set<Country> country = Countries.of("EE");
-    MatchResponse emptyResponse =
-        new MatchResponse(objectMapper.createArrayNode(), objectMapper.createObjectNode());
-    when(pepAndSanctionCheckService.match(user, country)).thenReturn(emptyResponse);
-    when(amlCheckRepository.findFirstByPersonalCodeAndTypeOrderByCreatedTimeDescIdDesc(
-            eq("123"), any(AmlCheckType.class)))
-        .thenReturn(Optional.empty());
-
-    assertFalse(amlService.isSanctionAndPepClear(user, country));
-  }
-
-  @Test
-  void recordedCitizenships_readsEveryCitizenshipFromTheLatestCustodyCheck() {
-    var person = new PersonImpl("61506150006", "Mari", "Maasikas");
-    given(
-            amlCheckRepository.findFirstByPersonalCodeAndTypeOrderByCreatedTimeDescIdDesc(
-                person.getPersonalCode(), CUSTODY_RIGHT))
-        .willReturn(
-            Optional.of(
-                AmlCheck.builder()
-                    .personalCode(person.getPersonalCode())
-                    .type(CUSTODY_RIGHT)
-                    .success(true)
-                    .metadata(Map.of("citizenship", "EE", "citizenships", List.of("EE", "RU")))
-                    .build()));
-
-    assertThat(amlService.recordedCitizenships(person)).isEqualTo(Countries.of("EE", "RU"));
-  }
-
-  @Test
-  void recordedCitizenships_fallsBackToTheSingleCitizenshipOlderChecksRecorded() {
-    var person = new PersonImpl("61506150006", "Mari", "Maasikas");
-    given(
-            amlCheckRepository.findFirstByPersonalCodeAndTypeOrderByCreatedTimeDescIdDesc(
-                person.getPersonalCode(), CUSTODY_RIGHT))
-        .willReturn(
-            Optional.of(
-                AmlCheck.builder()
-                    .personalCode(person.getPersonalCode())
-                    .type(CUSTODY_RIGHT)
-                    .success(true)
-                    .metadata(Map.of("citizenship", "EE"))
-                    .build()));
-
-    assertThat(amlService.recordedCitizenships(person)).isEqualTo(Countries.of("EE"));
-  }
-
-  @Test
-  void recordedCitizenships_isEmptyWhenTheCustodyCheckRecordedNoCitizenship() {
-    var person = new PersonImpl("61506150006", "Mari", "Maasikas");
-    given(
-            amlCheckRepository.findFirstByPersonalCodeAndTypeOrderByCreatedTimeDescIdDesc(
-                person.getPersonalCode(), CUSTODY_RIGHT))
-        .willReturn(
-            Optional.of(
-                AmlCheck.builder()
-                    .personalCode(person.getPersonalCode())
-                    .type(CUSTODY_RIGHT)
-                    .success(false)
-                    .metadata(Map.of("outcome", "NO_CUSTODY"))
-                    .build()));
-
-    assertThat(amlService.recordedCitizenships(person)).isEmpty();
-  }
-
-  @Test
-  void recordedCitizenships_isEmptyWhenThereIsNoCustodyCheck() {
-    var person = new PersonImpl("38812121215", "Jaan", "Tamm");
-    given(
-            amlCheckRepository.findFirstByPersonalCodeAndTypeOrderByCreatedTimeDescIdDesc(
-                person.getPersonalCode(), CUSTODY_RIGHT))
-        .willReturn(Optional.empty());
-
-    assertThat(amlService.recordedCitizenships(person)).isEmpty();
   }
 }
